% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_fct_wday_to_month.R
\name{wDay_to_month}
\alias{wDay_to_month}
\title{Function to convert wDay to month name}
\usage{
wDay_to_month(wDay)
}
\arguments{
\item{wDay}{numeric value of the water day}
}
\value{
month name
}
\description{
This function converts a wDay to a month name. Used as a global function for most plots
}
